% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ExonRSessionContext}
\alias{ExonRSessionContext}
\title{An ExonRSessionContext is a context for an Exon session.}
\description{
An ExonRSessionContext is a context for an Exon session.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExonRSessionContext-new}{\code{ExonRSessionContext$new()}}
\item \href{#method-ExonRSessionContext-sql}{\code{ExonRSessionContext$sql()}}
\item \href{#method-ExonRSessionContext-execute}{\code{ExonRSessionContext$execute()}}
\item \href{#method-ExonRSessionContext-clone}{\code{ExonRSessionContext$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExonRSessionContext-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExonRSessionContext-new}{}}}
\subsection{Method \code{new()}}{
Initialize the ExonRSessionContext object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExonRSessionContext$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExonRSessionContext-sql"></a>}}
\if{latex}{\out{\hypertarget{method-ExonRSessionContext-sql}{}}}
\subsection{Method \code{sql()}}{
Execute an SQL query.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExonRSessionContext$sql(query)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{query}}{The SQL query to execute. Returns an ExonDataFrame.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExonRSessionContext-execute"></a>}}
\if{latex}{\out{\hypertarget{method-ExonRSessionContext-execute}{}}}
\subsection{Method \code{execute()}}{
Execute an SQL query.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExonRSessionContext$execute(query)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{query}}{The SQL query to execute. Runs eagerly.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExonRSessionContext-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExonRSessionContext-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExonRSessionContext$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
