% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_exons.R
\name{find_exons}
\alias{find_exons}
\title{Title}
\usage{
find_exons(gene)
}
\arguments{
\item{gene}{is a string for the gene using RefSeq IDs}
}
\value{
a dataframe of the exons, and  their positions within the genome
}
\description{
Title
}
\examples{
e.g. Find the all the exons in BRCA2:
find_exons("NM_000059")

}
