% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_bed_file}
\alias{read_bed_file}
\title{Read BED File}
\usage{
read_bed_file(file_path)
}
\arguments{
\item{file_path}{The path to the BED file.}
}
\value{
A RecordBatch stream representing the contents of the BED table.
}
\description{
This function reads a BED file from a file and returns a RecordBatch
stream.
}
