% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_fasta_file}
\alias{read_fasta_file}
\title{Read FASTA File}
\usage{
read_fasta_file(file_path)
}
\arguments{
\item{file_path}{The path to the FASTA file.}
}
\value{
A RecordBatch stream representing the contents of the FASTA table.
}
\description{
This function reads a FASTA file from a file and returns a RecordBatch
stream.
}
