% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{read_fasta_file_extendr}
\alias{read_fasta_file_extendr}
\title{Return string \code{"Hello world!"} to R.}
\usage{
read_fasta_file_extendr(file_path, stream_ptr)
}
\description{
Return string \code{"Hello world!"} to R.
}
