% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_fastq_file}
\alias{read_fastq_file}
\title{Read FASTQ File}
\usage{
read_fastq_file(file_path)
}
\arguments{
\item{file_path}{The path to the FASTQ file.}
}
\value{
A RecordBatch stream representing the contents of the FASTQ table.
}
\description{
This function reads a FASTQ file from a file and returns a RecordBatch
stream.
}
