% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_genbank_file}
\alias{read_genbank_file}
\title{Read GenBank File}
\usage{
read_genbank_file(file_path)
}
\arguments{
\item{file_path}{The path to the GenBank file.}
}
\value{
A RecordBatch stream representing the contents of the GenBank table.
}
\description{
This function reads a GenBank file from a file and returns a RecordBatch
stream.
}
