% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_gff_file}
\alias{read_gff_file}
\title{Read GFF File}
\usage{
read_gff_file(file_path)
}
\arguments{
\item{file_path}{The path to the GFF file.}
}
\value{
A RecordBatch stream representing the contents of the GFF table.
}
\description{
This function reads a GFF file from a file and returns a RecordBatch
stream.
}
