% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{read_inferred_exon_table}
\alias{read_inferred_exon_table}
\title{Copy the inferred exon table from the given path into the given stream.}
\usage{
read_inferred_exon_table(file_path, stream_ptr)
}
\description{
Copy the inferred exon table from the given path into the given stream.
}
