% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_inferred_exon_table_r}
\alias{read_inferred_exon_table_r}
\title{Copy the inferred exon table from the given path into the given stream.}
\usage{
read_inferred_exon_table_r(file_path, pointer_addr)
}
\arguments{
\item{file_path}{The path to the inferred exon table.}

\item{stream_ptr}{The pointer to the stream to copy the inferred exon table}
}
\description{
Copy the inferred exon table from the given path into the given stream.
}
