% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_mzml_file}
\alias{read_mzml_file}
\title{Read MzML File}
\usage{
read_mzml_file(file_path)
}
\arguments{
\item{file_path}{The path to the MzML file.}
}
\value{
A RecordBatch stream representing the contents of the MzML file.
}
\description{
This function reads a MzML file from a file and returns a RecordBatch
stream.
}
