% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_sam_file}
\alias{read_sam_file}
\title{Read SAM/BAM File}
\usage{
read_sam_file(file_path)
}
\arguments{
\item{file_path}{The path to the SAM/BAM file.}
}
\value{
A RecordBatch stream representing the contents of the SAM/BAM file.
}
\description{
This function reads a SAM file from a file and returns a RecordBatch
stream. It will also read BAM files.
}
