% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_vcf_file}
\alias{read_vcf_file}
\title{Read VCF/BCF File}
\usage{
read_vcf_file(file_path)
}
\arguments{
\item{file_path}{The path to the VCF/BCF file.}
}
\value{
A RecordBatch stream representing the contents of the VCF/BCF file.
}
\description{
This function reads a VCF/BCF file from a file and returns a RecordBatch
stream.
}
